requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
numpy==1.26.4
opencv_python_headless==4.9.0.80
Pillow==10.2.0
rembg==2.0.53
setuptools==63.2.0
Werkzeug==3.0.1
Flask-Session==0.5.0
